requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsons==1.6.3
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
simplejson==3.19.3
typish==1.9.3
Werkzeug==3.1.3
